refactor: 전화번호 포매팅함수 유틸 모듈로 분리
