+++ frontend/src/hooks/useApplicantRegisterForm.js
@@ -1,16 +1,7 @@
 import { useState } from "react";
+import { formatHyphen } from "../utils/format/phoneNumber";
 
 const MAX_PHONE_NUMBER_LENGTH = 13;
-
-const setHyphen = (str, firstHyphenIdx, secondHyphenIdx) =>
-  str
-    .replaceAll("-", "")
-    .split("")
-    .map((el, idx) =>
-      idx === firstHyphenIdx || idx === secondHyphenIdx ? ["-", el] : el
-    )
-    .flat()
-    .join("");
 
 const useApplicantRegisterForm = () => {
   const [phoneNumber, setPhoneNumber] = useState("");
@@ -24,7 +15,7 @@ const useApplicantRegisterForm = () => {
     const firstHyphenIdx = value.length === MAX_PHONE_NUMBER_LENGTH ? 3 : 2;
     const secondHyphenIdx = value.length === MAX_PHONE_NUMBER_LENGTH ? 7 : 6;
 
-    const result = setHyphen(value, firstHyphenIdx, secondHyphenIdx).trim();
+    const result = formatHyphen(value, firstHyphenIdx, secondHyphenIdx).trim();
 
     setPhoneNumber(result);
   };
