+++ frontend/src/pages/PasswordFind/PasswordFind.js
@@ -9,7 +9,7 @@ import useForm from "../../hooks/useForm";
 import FormProvider from "../../provider/FormProvider/FormProvider";
 import InputField from "../../provider/FormProvider/InputField";
 import SubmitButton from "../../provider/FormProvider/SubmitButton";
-import { formatBirthday } from "../../utils/date";
+import { formatBirthday } from "../../utils/format/date";
 import {
   validateDay,
   validateMonth,
