+++ frontend/src/utils/format/phoneNumber.js
@@ -1,0 +1,9 @@
+export const formatHyphen = (str, firstHyphenIdx, secondHyphenIdx) =>
+  str
+    .replaceAll("-", "")
+    .split("")
+    .map((el, idx) =>
+      idx === firstHyphenIdx || idx === secondHyphenIdx ? ["-", el] : el
+    )
+    .flat()
+    .join("");
